requirements.txt
aws-cdk-lib==2.131.0
constructs>=10.0.0,<11.0.0
aws_solutions_constructs.aws_apigateway_lambda>=2.54.0